requirements.txt
flask
gunicorn
torch
torchvision
pillow
flask-cors